refactored notices
