+++ client/my-sites/stats/site.jsx
@@ -238,7 +238,7 @@ class StatsSite extends Component {
 					siteId={ siteId }
 					slug={ slug }
 				/>
-				<StatsNotices siteId={ siteId } />
+				<StatsNotices siteId={ siteId } isOdysseyStats={ isOdysseyStats } />
 				<HighlightsSection siteId={ siteId } currentPeriod={ defaultPeriod } />
 				<div id="my-stats-content" className={ wrapperClass }>
 					<>
