+++ client/my-sites/stats/stats-notices/feedback-notice.tsx
@@ -1,0 +1,47 @@
+import NoticeBanner from '@automattic/components/src/notice-banner';
+import { useTranslate } from 'i18n-calypso';
+
+interface FeedbackNoticeProps {
+	onTakeSurveyClick: () => void;
+	onRemindMeLaterClick: () => void;
+	onNoticeDismiss: () => void;
+}
+
+const FeedbackNotice = ( {
+	onTakeSurveyClick,
+	onRemindMeLaterClick,
+	onNoticeDismiss,
+}: FeedbackNoticeProps ) => {
+	const translate = useTranslate();
+
+	return (
+		<div className="inner-notice-container has-odyssey-stats-bg-color">
+			<NoticeBanner
+				level="info"
+				title={ translate( "We'd love to hear your thoughts on the new Stats" ) }
+				onClose={ onNoticeDismiss }
+			>
+				{ translate(
+					"{{p}}Now that you've gotten familiar with the new Jetpack Stats, we'd love to hear about your experience so we can continue to shape Jetpack to meet your needs.{{/p}}{{p}}{{takeSurveyButton}}Take quick survey{{/takeSurveyButton}} {{remindMeLaterLink}}Remind me later{{/remindMeLaterLink}}{{/p}}",
+					{
+						components: {
+							p: <p />,
+							takeSurveyButton: (
+								<button
+									type="button"
+									className="notice-banner__action-button"
+									onClick={ onTakeSurveyClick }
+								/>
+							),
+							remindMeLaterLink: (
+								<button className="notice-banner__action-link" onClick={ onRemindMeLaterClick } />
+							),
+						},
+					}
+				) }
+			</NoticeBanner>
+		</div>
+	);
+};
+
+export default FeedbackNotice;
