+++ client/my-sites/stats/stats-notices/opt-out-notice.tsx
@@ -1,0 +1,34 @@
+import NoticeBanner from '@automattic/components/src/notice-banner';
+import { useTranslate } from 'i18n-calypso';
+
+interface OptOutNoticeProps {
+	onNoticeDismiss: () => void;
+}
+
+const OptOutNotice = ( { onNoticeDismiss }: OptOutNoticeProps ) => {
+	const translate = useTranslate();
+
+	return (
+		<div className="inner-notice-container has-odyssey-stats-bg-color">
+			<NoticeBanner
+				level="success"
+				title={ translate( 'Welcome to the new Jetpack Stats!' ) }
+				onClose={ onNoticeDismiss }
+			>
+				{ translate(
+					'{{p}}Enjoy a more modern and mobile friendly experience with new stats and insights to help you grow your site.{{/p}}{{p}}If you prefer to continue using the traditional stats, {{manageYourSettingsLink}}manage your settings{{/manageYourSettingsLink}}.{{/p}}',
+					{
+						components: {
+							p: <p />,
+							manageYourSettingsLink: (
+								<a href="/wp-admin/admin.php?page=jetpack#/settings?term=stats" />
+							),
+						},
+					}
+				) }
+			</NoticeBanner>
+		</div>
+	);
+};
+
+export default OptOutNotice;
